+++ exporters/otlp/logs/src/testGrpcOkhttp/java/io/opentelemetry/exporter/otlp/logs/OtlpGrpcNettyOkHttpLogExporterTest.java
@@ -9,7 +9,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.grpc.inprocess.InProcessChannelBuilder;
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.exporter.internal.grpc.DefaultGrpcExporter;
+import io.opentelemetry.exporter.internal.grpc.UpstreamGrpcExporter;
 import io.opentelemetry.exporter.internal.marshal.Marshaler;
 import io.opentelemetry.exporter.internal.otlp.logs.ResourceLogsMarshaler;
 import io.opentelemetry.exporter.otlp.testing.internal.AbstractGrpcTelemetryExporterTest;
@@ -40,7 +40,7 @@ class OtlpGrpcNettyOkHttpLogExporterTest
         OtlpGrpcLogExporter.builder()
             .setChannel(InProcessChannelBuilder.forName("test").build())
             .build()) {
-      assertThat(exporter).extracting("delegate").isInstanceOf(DefaultGrpcExporter.class);
+      assertThat(exporter).extracting("delegate").isInstanceOf(UpstreamGrpcExporter.class);
     }
   }
